- allow the language table to supersede the title patches, if appropriate
For the Doom IWADs the provided font looks almost identical to the characters used on the title patches. So, for any level name that got replaced in some language, it will now check if the retrieved name comes from the default table, and if not, ignore the title patch and print the name with the specified font.

This also required removing the 'en' label from the default table,  because with this present, the text would always be picked from 'en' instead of 'default'. Since 'en' and 'default' had the same contents, in any English locale the 'default' table was never hit, so this won't make any difference for the texts being chosen.

Last but not least, wminfo has been made a local variable in G_DoCompleted. There were two places where this was accessed from outside the summary screen or its setup code, and both were incorrect.

+++ src/scripting/backend/codegen.cpp
@@ -322,7 +322,7 @@ static FxExpression *StringConstToChar(FxExpression *basex)
 	int chr = str.GetNextCharacter(position);
 
 	// Only succeed if the full string is consumed, i.e. it contains only one code point.
-	if (position == str.Len())
+	if (position == (int)str.Len())
 	{
 		return new FxConstant(chr, basex->ScriptPosition);
 	}
